# spec/openapi.yaml
openapi: '3.1.0'
info:
  title: Open API
  version: '1.0'
servers:
  - url: https://localhost:8080/api/v1
paths:
  /genres:
    get:
      operationId: getGenres
      tags:
        - genres
      summary: get music genres list
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreListResponse'
  /playlist/{id}:
    get:
      operationId: getPlaylist
      tags:
        - playlist
      summary: get playlist
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistResponse'
        "204":
          description: No content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /playlist:
    post:
      operationId: createPlaylist
      tags:
        - playlist
      summary: create playlist
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaylistRequestBody'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    GenreListResponse:
      required:
        - genreList
      title: GenreListResponse
      type: object
      properties:
        genres:
          type: array
          items:
            required:
            - genre
            properties:
              genre: 
                type: string
    
    PlaylistResponse:
      required:
      - id
      - playlist
      title: Playlist response
      type: object
      properties:
        id: 
          type: integer
        playlist:
          type: array
          items:
            $ref: '#/components/schemas/Music'
    
    CreatePlaylistRequestBody:
      required:
      - playlist
      title: Playlist response
      type: object
      properties:
        playlist:
          type: array
          items:
            $ref: '#/components/schemas/Music'
    
    Music:
      required:
      - genre
      - name
      - album
      - artist
      properties:
        genre:
          type: string
        name:
          type: string
        album:
          type: string
        artist:
          type: string

          
    Error:
      required:
        - timestamp
        - status
        - error
        - message
        - path
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        

                